Move wrapper class to new package
